% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_plot.R
\name{barcode_plot}
\alias{barcode_plot}
\title{Barcode plot}
\usage{
barcode_plot(
  data,
  protein_id,
  target,
  peptides,
  start,
  end,
  protein_length,
  foldchange,
  foldchange_cutoff = 1
)
}
\arguments{
\item{data}{Dataframe containing at least the input variables.}

\item{protein_id}{Column in the data dataframe containing the protein IDs.}

\item{target}{Protein ID for which a barcode plot should be plotted.}

\item{peptides}{Column containing the sequences.}

\item{start}{Start position of the peptide in the protein sequence.}

\item{end}{End position of the peptide in the protein sequence.}

\item{protein_length}{Length of the protein.}

\item{foldchange}{Fold change values for each peptide.}

\item{foldchange_cutoff}{Optional argument specifying the cutoff for highlighted peptides in the plot, default = 1.}
}
\value{
A barcode plot for the target protein.
}
\description{
Plots a "barcode plot" - a line for each identified peptide at the identified position with peptides changing in abundance highlighted.
}
\examples{
\dontrun{
barcode_plot(
data,
protein_id = uniprot_id,
target = "Q9Y6K9",
peptides = pep_stripped_sequence,
start = start, 
end = end,
protein_length = length_protein,
foldchange = Log2FC,
foldchange_cutoff = 0.5
)
}
}
