% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_peptide_abundance.R
\name{calculate_peptide_abundance}
\alias{calculate_peptide_abundance}
\title{Label-free peptide quantification}
\usage{
calculate_peptide_abundance(
  data,
  sample,
  peptide,
  precursor,
  intensity_log2,
  min_n_precursors = 1,
  method = "sum",
  for_plot = FALSE,
  retain_columns = NULL
)
}
\arguments{
\item{data}{a data frame that contains at least the input variables.}

\item{sample}{a character column in the \code{data} data frame that contains the sample name.}

\item{peptide}{a character column in the \code{data} data frame that contains a unique peptide
id, e.g. sequence.}

\item{precursor}{a character column in the \code{data} data frame that contains precursors.}

\item{intensity_log2}{a numeric column in the \code{data} data frame that contains log2
transformed precursor intensities.}

\item{min_n_precursors}{An integer specifying the minimum number of precursors required
for a peptide to be included in the analysis. The default value is 1, which means
no peptides are filtered.}

\item{method}{a character value specifying with which method peptide quantities should be
calculated. Possible options include \code{"sum"}, which takes the sum of all precursor
intensities as the peptide abundance. Another option is \code{"iq"}, which performs peptide
quantification based on a maximal peptide ratio extraction algorithm that is adapted from the
MaxLFQ algorithm of the MaxQuant software. Functions from the
\code{iq} package (\doi{10.1093/bioinformatics/btz961}) are used. Default is \code{"iq"}.}

\item{for_plot}{a logical value indicating whether the result should be only peptide intensities
or peptide intensities together with precursor intensities that can be used for plotting using
\code{peptide_profile_plot()}. Default is \code{FALSE}.}

\item{retain_columns}{a vector indicating if certain columns should be retained from the input
data frame. Default is not retaining additional columns \code{retain_columns = NULL}. Specific
columns can be retained by providing their names (not in quotations marks, just like other
column names, but in a vector).}
}
\value{
If \code{for_plot = FALSE}, protein abundances are returned, if \code{for_plot = TRUE}
also precursor intensities are returned in a data frame. The later output is ideal for plotting
with \code{peptide_profile_plot()} and can be filtered to only include protein abundances.
}
\description{
Determines relative peptide abundances from ion quantification.
By default peptides with a single precursor are considered.
If min precursor is set, the rule is applied to each peptide independently
}
\examples{
\donttest{
# Create example data
data <- data.frame(
  sample = c(
    rep("S1", 6),
    rep("S2", 6),
    rep("S1", 2),
    rep("S2", 2)
  ),
  peptide_id = c(
    rep(c("A", "A", "B", "B", "C", "D"), 2),
    rep(c("E", "F"), 2)
  ),
  precursor = c(
    rep(c("A1", "A2", "B1", "B2", "C1", "D1"), 2),
    rep(c("E1", "F1"), 2)
  ),
  intensity = c(
    rnorm(n = 2, mean = 18, sd = 1),
    rnorm(n = 2, mean = 21, sd = 1),
    rnorm(n = 1, mean = 21, sd = 1),
    rnorm(n = 1, mean = 15, sd = 1),
    rnorm(n = 2, mean = 13, sd = 1),
    rnorm(n = 2, mean = 15, sd = 1),
    rnorm(n = 1, mean = 15, sd = 1),
    rnorm(n = 1, mean = 18, sd = 1),
    rnorm(n = 1, mean = 12, sd = 1),
    rnorm(n = 1, mean = 13, sd = 1),
    rnorm(n = 1, mean = 15, sd = 1),
    rnorm(n = 1, mean = 16, sd = 1)
  )
)

data

# Calculate peptide abundances
peptide_abundance <- calculate_peptide_abundance(
  data,
  sample = sample,
  peptide = peptide_id,
  precursor = precursor,
  intensity_log2 = intensity,
  method = "sum",
  for_plot = FALSE
)

peptide_abundance

# Calculate peptide abundances and retain precursor
# abundances that can be used in a peptide profile plot
complete_abundances <- calculate_peptide_abundance(
  data,
  sample = sample,
  peptide = peptide_id,
  precursor = precursor,
  intensity_log2 = intensity,
  method = "sum",
  for_plot = TRUE
)

complete_abundances
}
}
