% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_protein_abundance.R
\name{calculate_protein_abundance}
\alias{calculate_protein_abundance}
\title{Label-free protein quantification}
\usage{
calculate_protein_abundance(
  data,
  sample,
  protein_id,
  precursor,
  intensity,
  method = "iq",
  for_plot = FALSE
)
}
\arguments{
\item{data}{A data frame that contains at least the input variables.}

\item{sample}{The name of the column containing the sample name.}

\item{protein_id}{The name of the column containing the protein accession numbers.}

\item{precursor}{The name of the column containing precursors.}

\item{intensity}{The name of the column containing log2 transformed precursor intensities.}

\item{method}{A character vector specifying with which method protein quantities should be calculated. Possible options include \code{"sum"},
which takes the sum of all precursor intensities as the protein abundance. Another option is \code{"iq"}, which performs protein 
quantification based on a maximal peptide ratio extraction algorithm that is adapted from the MaxQuant software. Functions from the \code{iq} 
package (https://academic.oup.com/bioinformatics/article/36/8/2611/5697917) are used. Default is \code{"iq"}.}

\item{for_plot}{A logical indicating whether the result should be only protein intensities or protein intensities together with precursor 
intensities that can be used for plotting using \code{qc_protein_abundance}. Default is \code{FALSE}.}
}
\value{
If \code{for_plot = FALSE}, protein abundances are returned, if \code{for_plot = TRUE} also precursor intensities are returned. The 
later output is ideal for plotting with \code{qc_protein_abundance} and can be filtered to only include protein abundances.
}
\description{
Determines relative protein abundances from ion quantification. Only proteins with at least 3 precursors are considered for quantification.
}
\examples{
\dontrun{
calculate_protein_abundance(
data,
sample = r_file_name,
protein_id = pg_protein_accessions,
precursor = eg_precursor_id,
intensity = log2_normalised_intensity,
method = "iq"
)
}
}
