% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_cv.R
\name{filter_cv}
\alias{filter_cv}
\title{Data filtering based on coefficients of variation (CV)}
\usage{
filter_cv(
  data,
  grouping,
  condition,
  log2_intensity,
  cv_limit = 0.25,
  min_conditions
)
}
\arguments{
\item{data}{Data frame containing at least the input variables.}

\item{grouping}{Column in the data frame containing the grouping variable that can be either precursors, peptides or proteins.}

\item{condition}{Column in the data frame containing information on the sample condition.}

\item{log2_intensity}{Column in the data frame containing log2 transformed intensities.}

\item{cv_limit}{Optional argument specifying the CV cutoff that will be applied. Default is 0.25.}

\item{min_conditions}{The minimum number of conditions for which grouping CVs should be below the cutoff.}
}
\value{
The CV filtered data frame.
}
\description{
Filters the input data based on precursor, peptide or protein intensity coefficients of variation.
}
\examples{
\dontrun{
filter_cv(
  data,
  grouping = eg_precursor_id,
  intensity = log2_intensity,
  condition = r_condition,
  log2_intensity = normalised_intensity_log2,
  cv_limit = 0.25,
  min_conditions = 5
)
}
}
