% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fmb_annotation_uniprot}
\alias{fmb_annotation_uniprot}
\title{List of manual annotations of feature metal binding sites}
\format{
A data.frame containing cleaned up feature metal binding sites and manual annotations with
ChEBI IDs.
}
\source{
UniProt ft_metal and ChEBI
}
\usage{
fmb_annotation_uniprot
}
\description{
A list that contains all feature metal binding site names that appear in UniProt. These sites
have been manually annotated with ChEBI IDs that match the name.
This was last updated on the 22/06/03.
}
\keyword{datasets}
