% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mako_colours}
\alias{mako_colours}
\title{Viridis colour scheme}
\format{
A vector containing 256 colours
}
\source{
created for the Seaborn statistical data visualization package for Python
}
\usage{
mako_colours
}
\description{
A perceptually uniform colour scheme originally created for the Seaborn python package.
}
\keyword{datasets}
