% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_normalisation.R
\name{median_normalisation}
\alias{median_normalisation}
\title{Median normalisation}
\usage{
median_normalisation(data, sample, grouping, log2intensity, na.rm = TRUE)
}
\arguments{
\item{data}{A data frame containing at least sample names, grouping variables and intensity values.}

\item{sample}{The name of the column containing the sample names.}

\item{grouping}{The name of the column containing the grouping variable - this can be peptides, precursors or proteins.}

\item{log2intensity}{The name of the column containing the log2 transformed intensity values to be normalised.}

\item{na.rm}{Logical indicating whether missing values should be removed. Default is TRUE.}
}
\value{
A new column in the original dataframe called \code{normalised_intensity_log2} containing the normalised intensity values.
}
\description{
Normalises the intensities in the way this is implemented in Spectronaut. The normalised intensity is the original intensity minus the run median plus the global median.
}
\examples{
\dontrun{
median_normalisation(data,
sample = r_file_name,
grouping = eg_precursor_id,
log2intensity = intensity_log2)
}
}
