% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metal_go_slim_subset}
\alias{metal_go_slim_subset}
\title{Molecular function gene ontology metal subset}
\format{
A data.frame containing a slim subset of molecular function gene ontology terms
that are related to metal binding. The \code{slims_from_id} column contains all IDs relevant
in this subset while the \code{slims_to_ids} column contains the starting IDs. If ChEBI IDs
have been annotated manually this is indicated in the \code{database} column.
}
\source{
QuickGO and ChEBI
}
\usage{
metal_go_slim_subset
}
\description{
A subset of molecular function gene ontology terms related to metals that was created
using the slimming process provided by the QuickGO EBI database.
This was last updated on the 19/02/24.
}
\keyword{datasets}
