% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metal_list}
\alias{metal_list}
\title{List of metals}
\format{
A data.frame containing the columns \code{atomic_number}, \code{symbol}, \code{name},
\code{type}, \code{chebi_id}, \code{chebi_ion_id}.
}
\source{
https://en.wikipedia.org/wiki/Metal and https://en.wikipedia.org/wiki/Metalloid
}
\usage{
metal_list
}
\description{
A list of all metals and metalloids in the periodic table.
}
\keyword{datasets}
