% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_fit_drc_4p.R
\name{parallel_fit_drc_4p}
\alias{parallel_fit_drc_4p}
\title{Fitting four-parameter dose response curves (using parallel processing)}
\usage{
parallel_fit_drc_4p(data, sample, grouping, response, dose, n_cores = NULL)
}
\arguments{
\item{data}{A data frame containing at least the input variables.}

\item{sample}{The name of the column containing the sample names.}

\item{grouping}{The name of the column containing precursor, peptide or protein identifiers.}

\item{response}{The name of the column containing response values, eg. log2 transformed intensities.}

\item{dose}{The name of the column containing dose values, eg. the treatment concentrations.}

\item{n_cores}{Optional, the number of cores used if workers are set up manually.}
}
\value{
A data frame is returned that contains correlations of predicted to measured values as a measure of the goodness of the curve fit, 
an associated p-value and the four parameters of the model for each group. Furthermore, input data for plots is returned in the columns \code{plot_curve} 
(curve and confidence interval) and \code{plot_points} (measured points).
}
\description{
This function is a wrapper around \code{fit_drc_4p} that allows the use of all system cores for model fitting. It should only be used on systems that have enough memory available. 
Workers can either be set up manually before running the function with \code{future::plan(multiprocess)} or automatically by the function. If workers are set up manually the 
number of cores should be provided to \code{n_cores}. Worker can be terminated after completion with \code{future::plan(sequential)}. It is not possible to export the 
individual fit objects when using this function as compared to the non parallel function as they are too large for efficient export from the workers.
}
\examples{
\dontrun{
parallel_fit_drc_4p(
data,
sample = r_file_name,
grouping = eg_precursor_id,
response = intensity,
dose = concentration
)
}
}
