% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_protti.R
\name{PCA_protti}
\alias{PCA_protti}
\title{Plot principal component analysis}
\usage{
PCA_protti(data, sample, grouping, intensity, condition)
}
\arguments{
\item{data}{A dataframe containing  sample names, peptide or precursor identifiers, corresponding intensities and a condition column indicating e.g. the treatment.}

\item{sample}{The column in the data dataframe containing the sample name.}

\item{grouping}{The column in the data dataframe containing either precursor or peptide identifiers.}

\item{intensity}{Column containing the corresponding intensity values for each peptide or precursor.}

\item{condition}{Column indicating the treatment or condition for each sample.}
}
\value{
A plotted principal component analysis showing PC1 and PC2
}
\description{
Plots a principal component analysis based on peptide or precursor intensities.
}
\examples{
\dontrun{
PCA_protti(
data,
sample = r_file_name,
grouping = pep_stripped_sequence,
intensity = normalised_intensity_log2,
condition = r_condition
)
}

}
