% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_peptide_type.R
\name{peptide_type}
\alias{peptide_type}
\title{Assign peptide type}
\usage{
peptide_type(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function was deprecated due to its name changing to \code{assign_peptide_type()}.
}
\keyword{internal}
