% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_peptide_profiles.R
\name{plot_peptide_profiles}
\alias{plot_peptide_profiles}
\title{Peptide abundance profile plot}
\usage{
plot_peptide_profiles(
  data,
  sample,
  peptide,
  intensity,
  grouping,
  targets,
  split_all = FALSE,
  protein_abundance_plot = FALSE
)
}
\arguments{
\item{data}{Data frame containing at least the input variables.}

\item{sample}{Column in the data frame containing sample names.}

\item{peptide}{Column in the data frame containing peptide or precursor names.}

\item{intensity}{Column in the data frame containing log2 transformed intensities.}

\item{grouping}{Column in the data frame containing groups by which the data should be split. This can be for example protein IDs.}

\item{targets}{Character vector specifying elements of the grouping column for which a plot should be returned.}

\item{split_all}{Logical, specifying if a plot for each element of the grouping column should be returned. The default is FALSE. Be careful
with this arguments as the number of plots can be quite high.}

\item{protein_abundance_plot}{Logical, if the input for this plot comes directly from \code{calculate_protein_abundance} this argument
can be set to \code{TRUE}. This displays all peptides in gray, while the protein abundance is displayed in green.}
}
\value{
A list of peptide profile plots.
}
\description{
Creates a plot of peptide abundances across samples. This is helpful to investigate effects of peptide and protein abundance changes
in different samples and conditions.
}
\examples{
\dontrun{
plot_peptide_abundance(
  data,
  sample = r_file_name,
  peptide = eg_precursor_id,
  intensity = log2_intensity,
  grouping = pg_protein_accessions,
  target = c("P03421")
)
}
}
