% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_abundance_normalisation.R
\name{protein_abundance_normalisation}
\alias{protein_abundance_normalisation}
\title{Perform normalization of LiP peptide data by protein abundance from tryptic controls}
\usage{
protein_abundance_normalisation(
  lip_peptides,
  tc_proteins,
  sample,
  grouping,
  condition,
  protein_id,
  peptide_intensity,
  protein_intensity,
  method,
  paired = FALSE,
  replicate_index = NULL,
  comparison = NULL,
  completeness = 0.7
)
}
\arguments{
\item{lip_peptides}{A data frame containing LiP peptide level data. The \code{sample, protein_id, grouping, condition, peptide_intensity} variables 
are required. Ideally the output of \code{assign_missingness} or \code{impute} is used for the method \code{"ref_vs_rest"}, since here also the variable 
\code{missingness} is required in contrast to method \code{"all_vs_all"}.}

\item{tc_proteins}{A data frame containing tryptic control protein level data. The \code{sample, protein_id, condition, protein_intensity} variables
are required. Ideally the output of \code{assign_missingness} or \code{impute} is used for the method \code{"ref_vs_rest"}, since here also the variable 
\code{missingness} is required in contrast to method \code{"all_vs_all"}.}

\item{sample}{The column in the data frame containing the sample name. Is not required if \code{method = "t-test_mean_sd"}.}

\item{grouping}{The column in the data frame containing precursor or peptide identifiers.}

\item{condition}{The column in the data frame containing the conditions.}

\item{protein_id}{The column in the data frame containing the protein accession numbers.}

\item{peptide_intensity}{The column in the data frame containing peptide intensity values. The intensity values need to be log2 transformed.}

\item{protein_intensity}{The column in the data frame containing protein intensity values. The intensity values need to be log2 transformed.}

\item{method}{A character vector specifying which method should be used. If \code{method = "all_vs_all"} an ANOVA based method will be performed that 
compares all conditions with each other followed by Tukey's honestly significant difference test to identify pairs of conditions that are significantly different.
If \code{method = "ref_vs_rest"} a reference condition will be compared against all other conditions using a Welch t-test in order to identify significantly
changing peptides or precursors.}

\item{paired}{A logical indicating if LiP and tryptic control replicates for each condition are paired. That means they belong together because they were 
taken from the exact same sample and then split in LiP and tryptic control conditions. If they are paired protein abundances are only used if there was a 
corresponding peptide detected in the LiP condition.}

\item{replicate_index}{The column in the data frame containing numbering for condition replicates. This column is only required if LiP and tryptic control samples are paired.
In that case the corresponding samples in both files should have the same numbering.}

\item{comparison}{The column in the data frame containing comparison information of treatment/reference condition pairs. Can be obtained by 
calling \code{assign_missingness}.}

\item{completeness}{The minimal degree of data completeness to be considered for evaluation. Value has to be between 0 and 1, default is 0.7. It is multiplied 
with the number of replicates and then adjusted downward. The resulting number is the minimal number of observations for each condition to be considered for
evaluation.}
}
\value{
A data frame that contains significance test statistics and differential abundances (\code{diff}) for each peptide or precursor and the 
corresponding comparison of conditions. If \code{method = "all_vs_all"} was used, ANOVA statistics are reported (\code{ms_group, ms_error, f, pval, adj_pval, n_conditions}). 
\code{ms_group} is the error within a group and \code{ms_error} is the error between groups. Also Tukey's honestly significant difference test statistics are 
reported (\code{df, q_value, pval_tukey}). If \code{method = "ref_vs_rest"} Welch t-test statistics are reported (\code{std_error, t_statistic, pval, adj_pval}).
}
\description{
LiP peptide intensities are normalised by protein abundances from tryptic controls to correct for protein fold changes due to treatment.
}
\examples{
\dontrun{
protein_abundance_normalisation(
lip_peptides = data_lip, 
tc_proteins = data_tc, 
sample = r_file_name,
protein_id = pg_protein_accessions, 
grouping = eg_precursor_id, 
condition = r_condition,
peptide_intensity = normalised_intensity_log2,
protein_intensity = protein_intensity_log2,
method = "all_vs_all")
}
}
