% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protti_colors}
\alias{protti_colors}
\title{Colour scheme for protti}
\format{
A vector containing 100 colours
}
\source{
Dina's imagination.
}
\usage{
protti_colors
}
\description{
A colour scheme for protti that contains 100 colours.
}
\keyword{datasets}
