% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_ids.R
\name{qc_ids}
\alias{qc_ids}
\title{Check number of precursor/preptide/protein IDs}
\usage{
qc_ids(data, sample, grouping, condition = NULL, title = "ID count per sample")
}
\arguments{
\item{data}{Dataframe containing at least sample names and precursor IDs.}

\item{sample}{Column in the dataframe specifying the sample name.}

\item{grouping}{Column in the data dataframe containing either precursor or peptide identifiers.}

\item{condition}{Optional column in the dataframe specifying the condition of the sample (e.g. LiP_reated, LiP_untreated).}

\item{title}{Optional argument specifying the plot title (default is "ID count per sample").}
}
\value{
A bar plot with the height corresponding to the number of IDs, each bar represents one sample.
}
\description{
Plots the number of IDs for each sample
}
\examples{
\dontrun{
qc_ids(
data,
sample = r_file_name
grouping = eg_precursor_id,
condition = r_condition,
title = "Number of peptide IDs per sample"
)
}
}
