% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_log2_intensity_distribution.R
\name{qc_log2_intensity_distribution}
\alias{qc_log2_intensity_distribution}
\title{Check Log2 intensity distribution}
\usage{
qc_log2_intensity_distribution(data, sample, grouping, log2_intensity)
}
\arguments{
\item{data}{A dataframe containing at least sample names, peptide or precursor identifiers and log2 transformed intensities for each peptide or precursor.}

\item{sample}{The column in the data dataframe containing the sample name.}

\item{grouping}{The column in the data dataframe containing either precursor or peptide identifiers.}

\item{log2_intensity}{The column in the data dataframe containing the log2 transformed intensities of each precursor or peptide.}
}
\value{
A plot that shows the intensity distribution over all samples
}
\description{
Plots the overall distribution of all peptide intensities
}
\examples{
\dontrun{
qc_log2_intensity_distribution(
data,
sample = r_file_name,
grouping = pep_stripped_sequence,
log2_intensity = normalised_intensity_log2)
}
}
