% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_ranked_intensities.R
\name{qc_ranked_intensities}
\alias{qc_ranked_intensities}
\title{Ranked intensities}
\usage{
qc_ranked_intensities(
  data,
  sample,
  grouping,
  log2_intensity,
  facet = FALSE,
  interactive = FALSE
)
}
\arguments{
\item{data}{a data frame that contains at least the input variables.}

\item{sample}{a character column in the \code{data} data frame that contains the sample names.}

\item{grouping}{a character column in the \code{data} data frame that contains protein, precursor,
or peptide identifiers.}

\item{log2_intensity}{a numeric column in the \code{data} data frame that contains the log2
transfromed intensities of the selected grouping variable.}

\item{facet}{a logical value that specifies whether the plot should be faceted by samples
(default is FALSE).If \code{facet = FALSE} the median of each protein intensity will be shown in the plot.}

\item{interactive}{a logical value that specifies whether the plot should be interactive
(default is FALSE).}
}
\value{
A plot showing the ranked intensities is returned.
}
\description{
Plots ranked intensities
}
\examples{
set.seed(123) # Makes example reproducible

# Create synthetic data
data <- create_synthetic_data(
  n_proteins = 10,
  frac_change = 0.5,
  n_replicates = 4,
  n_conditions = 3,
  method = "effect_random",
  additional_metadata = FALSE
)

# Calculate protein abundance
protein_intensities <- calculate_protein_abundance(
                           data = data,
                           sample = sample, 
                           protein_id = protein, 
                           precursor = peptide,
                           peptide = peptide, 
                           intensity_log2 = peptide_intensity)
                          
# Plot ranked intensities for all samples combined
qc_ranked_intensities(data = protein_intensities, 
                     sample = sample, 
                     grouping = protein, 
                     log2_intensity = peptide_intensity, 
                     interactive = TRUE)

# Plot ranked intensities for each sample separately
qc_ranked_intensities(data = protein_intensities, 
                     sample = sample, 
                     grouping = protein, 
                     log2_intensity = peptide_intensity, 
                     interactive = TRUE, 
                     facet = TRUE)

}
