% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_run_intensity.R
\name{qc_run_intensity}
\alias{qc_run_intensity}
\title{Check run intensities of all samples}
\usage{
qc_run_intensity(data, sample, grouping, log2intensity)
}
\arguments{
\item{data}{A dataframe containing at least sample names, peptide or precursor identifiers and log2 transformed intensities for each peptide or precursor.}

\item{sample}{The column in the data dataframe containing the sample name.}

\item{grouping}{The column in the data dataframe containing either precursor or peptide identifiers.}

\item{log2intensity}{The column in the data dataframe containing the log2 transformed intensities of each precursor or peptide.}
}
\value{
A plot with one boxplot for each sample, depicting the peptide or precurosor intensities.
}
\description{
Plots the intensities in each sample as a boxplot
}
\examples{
\dontrun{
qc_run_intensity(data,
sample,
grouping,
log2intensity)
}
}
