% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_sequence_coverage.R
\name{qc_sequence_coverage}
\alias{qc_sequence_coverage}
\title{Protein coverage distribution}
\usage{
qc_sequence_coverage(
  data,
  protein_identifier,
  coverage,
  facet_by_sample = FALSE,
  sample = NULL,
  interactive = TRUE
)
}
\arguments{
\item{data}{A data frame containing at least the input variables.}

\item{protein_identifier}{Column in the data frame containing protein identifiers.}

\item{coverage}{Column in the data frame containing protein coverage in percent. This information can be obtained using 
the \code{\link{sequence_coverage}} function.}

\item{facet_by_sample}{Logical that determines if the plot should be faceted by sample. If \code{TRUE}, \code{sample} needs to be 
provided. Default is \code{FALSE}.}

\item{sample}{Column in the data frame containing sample names. Only required if \code{facet_by_sample = TRUE}.}

\item{interactive}{A logical, if TRUE the plot is interactive (default is TRUE).}
}
\value{
A protein coverage histogram with 5 percent binwidth. The vertical dotted line indicates the median.
}
\description{
Plots the distribution of protein coverages in a histogram.
}
\examples{
\dontrun{
qc_sequence_coverage(
data,
protein_identifier = pg_protein_accessions,
coverage = coverage
)
}

}
\seealso{
\code{\link{sequence_coverage}}
}
