% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rapamycin_10uM}
\alias{rapamycin_10uM}
\title{Rapamycin 10 uM example data}
\format{
A data frame containing peptide level data from a Spectronaut report.
}
\source{
Piazza, I., Beaton, N., Bruderer, R. et al. A machine learning-based chemoproteomic
approach to identify drug targets and binding sites in complex proteomes. Nat Commun 11, 4200
(2020). https://doi.org/10.1038/s41467-020-18071-x
}
\usage{
rapamycin_10uM
}
\description{
Rapamycin example data used for the vignette about binary control/treated data. The data was
obtained from \href{https://www.nature.com/articles/s41467-020-18071-x}{Piazza 2020}
and corresponds to experiment 18. FKBP1A the rapamycin binding protein and 49 other randomly
sampled proteins were used for this example dataset. Furthermore, only the DMSO control and the
10 uM condition were used.
}
\keyword{datasets}
