% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rapamycin_dose_response}
\alias{rapamycin_dose_response}
\title{Rapamycin dose response example data}
\format{
A data frame containing peptide level data from a Spectronaut report.
}
\source{
Piazza, I., Beaton, N., Bruderer, R. et al. A machine learning-based chemoproteomic approach to identify drug targets and binding sites in complex proteomes. Nat Commun 11, 4200 (2020). https://doi.org/10.1038/s41467-020-18071-x
}
\usage{
rapamycin_dose_response
}
\description{
Rapamycin example data used for the vignette about dose response data. The data is obtained from \href{https://www.nature.com/articles/s41467-020-18071-x}{Piazza 2020}
and corresponds to experiment 18. FKBP1A the rapamycin binding protein and 39 other randomly sampled proteins were used 
for this example dataset. The concentration range includes the following points: 0 (DMSO control), 10 pM, 100 pM, 1 nM, 10 nM, 100 nM, 1 uM, 10 uM and 100 uM.
}
\keyword{datasets}
