% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_identified_by_x.R
\name{replace_identified_by_x}
\alias{replace_identified_by_x}
\title{Replace each identified positions in protein by "x"}
\usage{
replace_identified_by_x(sequence, positions_start, positions_end)
}
\arguments{
\item{sequence}{Protein sequence}

\item{positions_start}{Start position of the identified peptide.}

\item{positions_end}{End position of the identified peptide.}
}
\value{
A modified protein sequence with each identified position replaced by "x"
}
\description{
Helper function for the calculation of sequence coverage, replaces identified positions with an "x"
}
\examples{
\dontrun{
replace_identified_by_x(
sequence = protein_sequence,
positions_start = start,
positions_end = end
)
}

}
