% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale}
\alias{scale}
\title{Scaling a vector}
\usage{
scale(x, method)
}
\arguments{
\item{x}{a numeric vector}

\item{method}{the method to be used for scaling. "01" scales the vector between 0 and 1. "center" scales the vector equal to \code{base::scale} around a center. This is done by subtracting the mean from every value and then deviding them by the standard deviation.}
}
\value{
A scaled numeric vector.
}
\description{
\code{scale} is used to scale a numeric vector either between 0 and 1 or around a centered value using the standard deviation.
}
\examples{
\dontrun{
scale(c(1, 2, 1, 4, 6, 8), method = "01")
}
}
