% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_metal_name.R
\name{split_metal_name}
\alias{split_metal_name}
\title{Convert metal names to search pattern}
\usage{
split_metal_name(metal_names)
}
\arguments{
\item{metal_names}{A character vector containing names of metals and metal containing molecules.}
}
\value{
A character vector with metal name search patterns.
}
\description{
Converts a vector of metal names extracted from the \code{feature_metal_binding} column obtained with \code{fetch_uniprot} to a pattern that can be used to search for corresponding ChEBI IDs. This is used as a helper function for other functions.
}
\examples{
\dontrun{
split_metal_name(
metal_names = c(
"Magnesium",
NA,
NA,
"Magnesium or Manganese",
"Divalent metal cation",
"Cobalt (cobalamin axial ligand)",
"Iron-sulfur (2Fe-2S)")
)
}
}
