% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_treatment_enrichment.R
\name{treatment_enrichment}
\alias{treatment_enrichment}
\title{Check treatment enrichment}
\usage{
treatment_enrichment(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function was deprecated due to its name changing to \code{calculate_treatment_enrichment()}.
}
\keyword{internal}
