% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest_protti.R
\name{ttest_protti}
\alias{ttest_protti}
\title{Perform Welch's t-test}
\usage{
ttest_protti(mean1, mean2, stddev1, stddev2, n1, n2)
}
\arguments{
\item{mean1}{Vector containing the means of group1.}

\item{mean2}{Vector containing the means of group2.}

\item{stddev1}{Vector containing the standard deviations of group1.}

\item{stddev2}{Vector containing the standard deviations of group2.}

\item{n1}{Vector containing the number of replicates used for the calculation of each mean and standard deviation of group1.}

\item{n2}{Vector containing the number of replicates used for the calculation of each mean and standard deviation of group2.}
}
\value{
A data frame that contains the calculated differences of means, standard error, t statistic and p-values.
}
\description{
Performs a Welch's t-test and calculates p-values between two groups.
}
\examples{
\dontrun{
ttest_protti(
mean1 = mean_untreated,
mean2 = mean_treated,
stddev1 = sd_untreated,
stddev2 = sd_treated,
n1 = count_untreated,
n2 = count_treated)
}
}
