% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_protti}
\alias{volcano_protti}
\title{Volcano plot}
\usage{
volcano_protti(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function was deprecated due to its name changing to \code{volcano_plot()}.
}
\keyword{internal}
