% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_missingness.R
\name{assign_missingness}
\alias{assign_missingness}
\title{Assignment of missingness types}
\usage{
assign_missingness(
  data,
  sample,
  condition,
  grouping,
  intensity,
  noise = NULL,
  ref_condition = "control"
)
}
\arguments{
\item{data}{a data frame containing at least the input variables.}

\item{sample}{the name of the column containing the sample names.}

\item{condition}{the name of the column containing the conditions.}

\item{grouping}{the name of the column containing precursor or peptide identifiers.}

\item{intensity}{the name of the column containing intensity values.}

\item{noise}{optional, the name of the column containing noise values for imputation. Only needed if noise imputation is performed.}

\item{ref_condition}{the condition that is used as a reference for missingness determination. By default \code{ref_condition = "control"}.}
}
\value{
A data frame that contains the reference condition paired with each treatment condition. The \code{comparison} column contains the comparison name for the specific reference/treatment pair. The \code{missingness} column reports the type of missingness. 
\itemize{
\item{"complete": }{No missing values for every replicate of this reference/treatment pair for the specific grouping variable.}
\item{"MNAR": }{Missing not at random. All replicates of either the reference or treatment condition have missing values for the specific grouping variable.}
\item{"MAR": }{Missing at random. At least n-1 replicates have missing values for the reference/treatment pair for the specific grouping varible.}
}
}
\description{
The type of missingness (missing at random, missing not at random) is assigned based on the comparison of a reference condition and every other condition.
}
\examples{
\dontrun{
assign_missingness(
data,
sample = r_file_name,
condition = r_condition,
grouping = eg_precursor_id,
intensity = normalised_intensity_log2
)
}
}
