% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_diff_abundance.R
\name{diff_abundance}
\alias{diff_abundance}
\title{Calculate differential abundance between conditions}
\usage{
diff_abundance(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function was deprecated due to its name changing to \code{calculate_diff_abundance()}.
}
\keyword{internal}
