% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_metal_binders.R
\name{extract_metal_binders}
\alias{extract_metal_binders}
\title{Extract metal-bind protein information from UniProt}
\usage{
extract_metal_binders(
  data,
  protein_id = id,
  feature_metal_binding = feature_metal_binding,
  chebi_cofactor = chebi_cofactor,
  chebi_catalytic_activity = chebi_catalytic_activity,
  comment_cofactor = comment_cofactor,
  chebi_data = NULL,
  chebi_relation_data = NULL
)
}
\arguments{
\item{data}{a data frame containing at least the input columns.}

\item{protein_id}{a character column in the \code{data} data frame that contains the protein
identifiers.}

\item{feature_metal_binding}{a character column in the \code{data} data frame that contains the
feature metal binding information from UniProt.}

\item{chebi_cofactor}{a character column in the \code{data} data frame that contains the ChEBI
cofactor information from UniProt.}

\item{chebi_catalytic_activity}{a character column in the \code{data} data frame that contains
the ChEBI catalytic activity information from UniProt.}

\item{comment_cofactor}{a character cholumn in the \code{data} data frame that contains
the comment cofactor information from UniProt.}

\item{chebi_data}{optional, a data frame that can be manually obtained with \code{fetch_chebi()}.
If not provided it will be fetched within the function. If the function is run many times it
is recommended to provide the data frame to save time.}

\item{chebi_relation_data}{optional, a data frame that can be manually obtained with
\code{fetch_chebi(relation = TRUE)}. If not provided it will be fetched within the function.
If the function is run many times it is recommended to provide the data frame to save time.}
}
\value{
A data frame containing information on protein metal binding state. It contains the
following types of columns (the naming might vary based on the input):
\itemize{
\item{\code{protein_id}: }{UniProt protein identifier.}
\item{\code{metal_type}: }{Metal name extracted from \code{feature_metal_binding} information.
This is the name that is used as a search pattern in order to assign a ChEBI ID with the
\code{split_metal_name} helper function within this function.}
\item{\code{metal_position}: }{Amino acid position within the protein that is involved in metal
binding.}
\item{\code{metal_id}: }{This is a unique identifier for the metal of a protein. If the protein
binds only one metal this will be 1. If the protein e.g. binds 2 magnesium ions, one will have
the ID 1, while the other has the ID 2.}
\item{\code{ids}: }{ChEBI ID assigned to protein and binding site based on \code{metal_type}
column name. These are general IDs that have sub-IDs. Thus, they generally describe the type of
metal ion bound to the protein.}
\item{\code{main_id_name}: }{Official ChEBI name associated with the ID in the \code{ids} column.}
\item{\code{sub_ids}: }{ChEBI ID that is a sub-ID (incoming) of the ID in the \code{ids} column.
Thus, they more specifically describe the potential nature of the metal ion.}
\item{\code{sub_id_name}: }{Official ChEBI name associated with the ID in the \code{sub_ids}
column.}
\item{\code{source}: }{The source of the information, can be either \code{feature_metal_binding},
\code{chebi_cofactor}, \code{comment_cofactor} or \code{chebi_catalytic_activity}.}
\item{\code{evidence}: }{Evidence level of this annotation. This is usually an evidence & conclusion
ontology ID (ECO ID). Sometimes a publication ID is included.}
\item{\code{note}: }{A note specifying the annotation further. Can for example specify how many
metals are bound per protein.}
}
}
\description{
Information of metal binding proteins is extracted from UniProt data retrieved with
\code{fetch_uniprot}. ChEBI IDs, potential sub-IDs for metal cations, binding site locations
in the protein and sub-ID evidence level (based on metal presence as cofactor) are extracted.
}
\examples{
\donttest{
# Create example data
data <- fetch_uniprot(
  uniprot_ids = c("Q03640", "Q03778", "P22276"),
  columns = c(
    "feature(METAL BINDING)",
    "chebi(Cofactor)",
    "chebi(Catalytic activity)",
    "comment(COFACTOR)"
  )
)

# Extract metal binding information
metal_info <- extract_metal_binders(
  data = data,
  protein_id = id,
  feature_metal_binding = feature_metal_binding,
  chebi_cofactor = chebi_cofactor,
  chebi_catalytic_activity = chebi_catalytic_activity,
  comment_cofactor = comment_cofactor
)

metal_info
}
}
