% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_chebi.R
\name{fetch_chebi}
\alias{fetch_chebi}
\title{Fetch ChEBI database information}
\usage{
fetch_chebi(relation = FALSE)
}
\arguments{
\item{relation}{a logical, if TRUE, ChEBI Ontology data will be returned instead the main compound data. This data can be used to check the relations of ChEBI ID's to eachother.}
}
\value{
A data frame that contains all information about each molecule in the ChEBI database. Only "3-star" observations are included in the result. These are entries manually annotated by the ChEBI curator team.
}
\description{
Fetches all information from the ChEBI database.
}
\examples{
\dontrun{
fetch_chebi()
}
}
