% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_eco.R
\name{fetch_eco}
\alias{fetch_eco}
\title{Fetch evidence & conclusion ontology}
\usage{
fetch_eco(
  return_relation = FALSE,
  return_history = FALSE,
  show_progress = TRUE
)
}
\arguments{
\item{return_relation}{a logical value that indicates if relational information should be returned instead
the main descriptive information. This data can be used to check the relations of ECO terms to each other.
Default is FALSE.}

\item{return_history}{a logical value that indicates if the entry history of an ECO term should be
returned instead the main descriptive information.
Default is FALSE.}

\item{show_progress}{a logical value that indicates if a progress bar will be shown.
Default is TRUE.}
}
\value{
A data frame that contains descriptive information about each ECO term in the EBI database.
If either \code{return_relation} or \code{return_history} is set to \code{TRUE}, the respective information is
returned instead of the usual output.
}
\description{
Fetches all evidence & conclusion ontology (ECO) information from the EBI database.
}
\examples{
\donttest{
eco <- fetch_eco()

head(eco)
}
}
