% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_pdb.R
\name{fetch_pdb}
\alias{fetch_pdb}
\title{Fetch protein structure data from RCSB}
\usage{
fetch_pdb(
  pdb_ids,
  columns = c("compound", "structureId", "entityId", "structureTitle", "chainId",
    "chainLength", "sequence", "classification", "db_name", "ligandName", "InChIKey",
    "uniprotAcc", "taxonomy", "taxonomyId", "experimentalTechnique", "resolution",
    "crystallizationMethod", "crystallizationTempK", "phValue", "method", "temperature",
    "ionicStrength", "ph"),
  batchsize = 400,
  show_progress = TRUE
)
}
\arguments{
\item{pdb_ids}{A character vector of pdb identifiers}

\item{columns}{A character vector of data columns that should be imported from RCSB (all possible columns can be found here: https://www.rcsb.org/pdb/results/reportField.do)}

\item{batchsize}{Size of batch of proteins for a single query}

\item{show_progress}{Logical, if true, a progress bar will be shown}
}
\value{
A data frame that contains all structural data specified in \code{columns} for the pdb structures provided.
}
\description{
Fetches protein structural data from the RCSB database.
}
\examples{
\dontrun{
fetch_pdb(c("3OFN", "3ZIA", "3ZRY"))
}
}
