% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_pdb_structure.R
\name{fetch_pdb_structure}
\alias{fetch_pdb_structure}
\title{Fetch PDB structure atom data from RCSB}
\usage{
fetch_pdb_structure(pdb_ids, return_data_frame = FALSE, show_progress = TRUE)
}
\arguments{
\item{pdb_ids}{a character vector of PDB identifiers.}

\item{return_data_frame}{logical, if true, a data frame instead of a list is returned. It is recommended to only use this if not many 
pdb structures are retrieved. Default is FALSE.}

\item{show_progress}{logical, if true, a progress bar will be shown. Default is TRUE.}
}
\value{
A list that contains atom data for each PDB structures provided. If return_data_frame is TRUE, a data frame with this
information is returned instead.
}
\description{
Fetches atom data for a PDB structure from RCSB. If you want to retrieve metadata about PDB structures, use the function \code{fetch_pdb()}.
}
\examples{
\donttest{
head(fetch_pdb_structure(pdb_ids = c("6HG1", "1E9I", "6D3Q", "4JHW"), return_data_frame = TRUE))
}
}
