% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_drc_4p.R
\name{fit_drc_4p}
\alias{fit_drc_4p}
\title{Fitting four-parameter dose response curves}
\usage{
fit_drc_4p(
  data,
  sample,
  grouping,
  response,
  dose,
  log_logarithmic = TRUE,
  include_models = FALSE,
  retain_columns = NULL
)
}
\arguments{
\item{data}{A data frame containing at least the input variables.}

\item{sample}{The name of the column containing the sample names.}

\item{grouping}{The name of the column containing precursor, peptide or protein identifiers.}

\item{response}{The name of the column containing response values, eg. log2 transformed intensities.}

\item{dose}{The name of the column containing dose values, eg. the treatment concentrations.}

\item{log_logarithmic}{logical indicating if a logarithmic or log-logarithmic model is fitted. 
If response values form a symmetric curve for non-log transformed dose values, a logarithmic model instead
of a log-logarithmic model should be used. Usually biological dose response data has a log-logarithmic distribution, which is the 
reason this is the default. Log-logarithmic models are symmetric if dose values are log transformed.}

\item{include_models}{A logical indicating if model fit objects should be exported. These are usually very large and not necessary for further analysis.}

\item{retain_columns}{A vector indicating if certain columns should be retained from the input data frame. Default is not retaining 
additional columns \code{retain_columns = NULL}. Specific columns can be retained by providing their names (not in quotations marks, 
just like other column names, but in a vector).}
}
\value{
If \code{include_models = FALSE} a data frame is returned that contains correlations of predicted to measured values as a measure of the goodness of the curve fit,
an associated p-value and the four parameters of the model for each group. Furthermore, input data for plots is returned in the columns \code{plot_curve} (curve and confidence interval) and \code{plot_points} (measured points).
If \ code{include_models = TURE}, a list is returned that contains: 
\itemize{
\item{\code{fit_objects}: }{The fit objects of type \code{drc} for each group.}
\item{\code{correlations}: }{The correlation data frame described above}
}
}
\description{
Function for fitting four-parameter dose response curves for each group (precursor, peptide or protein).
}
\examples{
\dontrun{
fit_drc_4p(
data,
sample = r_file_name,
grouping = eg_precursor_id,
response = intensity,
dose = concentration
)
}
}
