% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_drc_4p.R
\name{fit_drc_4p}
\alias{fit_drc_4p}
\title{Fitting four-parameter dose response curves}
\usage{
fit_drc_4p(data, sample, grouping, response, dose)
}
\arguments{
\item{data}{a data frame containing at least the input variables.}

\item{sample}{the name of the column containing the sample names.}

\item{grouping}{the name of the column containing precursor, peptide or protein identifiers.}

\item{response}{the name of the column containing response values, eg. log2 transformed intensities.}

\item{dose}{the name of the column containing dose values, eg. the treatment concentrations.}
}
\value{
A list that contains: 
\itemize{
\item{\code{fit_objects}: }{The fit objects of type \code{drc} for each group. }
\item{\code{correlations}: }{A data frame that contains correlation of predicted to measured values as a measure of the goodness of the curve fit, an associated p-value and the four parameters of the model for each group.}
\item{\code{plots}: }{The ggplot objects containing the curve plot. Printing these will print the plot.}
}
}
\description{
Function for fitting four-parameter dose response curves for each group (precursor, peptide or protein).
}
\examples{
\dontrun{
fit_drc_4p(
data,
sample = r_file_name,
grouping = eg_precursor_id,
response = intensity,
dose = concentration
)
}
}
