% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_go_enrichment.R
\name{go_enrichment}
\alias{go_enrichment}
\title{Perform gene ontology enrichment analysis}
\usage{
go_enrichment(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function was deprecated due to its name changing to \code{calculate_go_enrichment()}.
}
\keyword{internal}
