% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/go_enrichment.R
\name{go_enrichment}
\alias{go_enrichment}
\title{Perform gene ontology enrichment analysis}
\usage{
go_enrichment(data, protein_id, is_significant, go_annotations, plot = TRUE)
}
\arguments{
\item{data}{A data frame that contains at least the input variables.}

\item{protein_id}{The name of the column containing the protein accession numbers.}

\item{is_significant}{The name of the column containing a logical indicating if the corresponding protein has a significantly changing peptide. 
The input data frame may contain peptide level information with significance information. The function is able to extract protein level information from this.}

\item{go_annotations}{The name of the column containing gene ontology annotations. These can be obtained from UniProt using \code{fetch_uniprot}.}

\item{plot}{A logical indicating whether the result should be plotted or returned as a table.}
}
\value{
A bar plot displaying negative log10 adjusted p-values for the top 10 enriched gene ontology terms. Bars are colored according to the direction of the 
enrichment. If \code{plot = FALSE}, a data frame is returned.
}
\description{
Analyses enrichment of gene ontology terms associated with proteins in the fraction of significant proteins compared to all detected proteins. A Fisher's 
exact test is performed to test significance of enrichment.
}
\examples{
\dontrun{
go_enrichment(
data,
protein_id = pg_protein_accessions,
is_significant = significant,
go_annotations = go_molecular_function
)
}
}
