% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_enrichment.R
\name{kegg_enrichment}
\alias{kegg_enrichment}
\title{Perform KEGG pathway enrichment analysis}
\usage{
kegg_enrichment(
  data,
  protein_id,
  is_significant,
  pathway_id,
  pathway_name,
  plot = TRUE
)
}
\arguments{
\item{data}{A data frame that contains at least the input variables.}

\item{protein_id}{The name of the column containing the protein accession numbers.}

\item{is_significant}{The name of the column containing a logical indicating if the corresponding protein has a significantly changing peptide. 
The input data frame may contain peptide level information with significance information. The function is able to extract protein level information from this.}

\item{pathway_id}{The name of the column containing KEGG pathway identifiers. These can be obtained from KEGG using \code{fetch_kegg}.}

\item{pathway_name}{The name of the column containing KEGG pathway names. These can be obtained from KEGG using \code{fetch_kegg}.}

\item{plot}{A logical indicating whether the result should be plotted or returned as a table.}
}
\value{
A bar plot displaying negative log10 adjusted p-values for the top 10 enriched pathways. Bars are colored according to the direction of the 
enrichment. If \code{plot = FALSE}, a data frame is returned.
}
\description{
Analyses enrichment of KEGG pathways associated with proteins in the fraction of significant proteins compared to all detected proteins. A Fisher's 
exact test is performed to test significance of enrichment.
}
\examples{
\dontrun{
kegg_enrichment(
data,
protein_id = pg_protein_accessions,
is_significant = significant,
pathway_id = pathway_id,
pathway_name = pathway_name
)
}
}
