% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{median_normalisation}
\alias{median_normalisation}
\title{Intensity normalisation}
\usage{
median_normalisation(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function was deprecated due to its name changing to \code{normalise()}.
The normalisation method in the new function needs to be provided as an argument.
}
\keyword{internal}
