% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_normalisation.R
\name{median_normalisation}
\alias{median_normalisation}
\title{Median normalisation}
\usage{
median_normalisation(
  data,
  file_column_name,
  intensity_column_name,
  na.rm = FALSE
)
}
\arguments{
\item{data}{A data frame containing at least sample names and intensity values.}

\item{file_column_name}{The name of the column containing the sample names.}

\item{intensity_column_name}{The name of the column containing the intensity values.}

\item{na.rm}{A logical indicating wther missing values should be removed.}
}
\value{
A data frame that contains the input data and an additional column with normalised intensity values.
}
\description{
Median normalise intensity values for all samples of mass spectrometry data. The intensity values are ideally log2 transformed. The output is a new column called normalised_intensity_log2.
}
\examples{
\dontrun{
median_normalisation(data, sample_name, intensity_log2, na.rm = TRUE)
}
}
