% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_functional_network.R
\name{network_analysis}
\alias{network_analysis}
\title{Analyse protein interaction network for significant hits}
\usage{
network_analysis(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function was deprecated due to its name changing to \code{analyse_functional_network()}.
}
\keyword{internal}
