% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drc_4p.R
\name{plot_drc_4p}
\alias{plot_drc_4p}
\title{plotting four-parameter dose response curves}
\usage{
plot_drc_4p(
  data,
  grouping,
  response,
  dose,
  targets,
  unit = "uM",
  y_axis_name = "Response",
  scales = "free",
  x_axis_scale_log10 = TRUE
)
}
\arguments{
\item{data}{A data frame that is obtained by calling the \code{fit_drc_4p} function.}

\item{grouping}{The name of the column containing precursor, peptide or protein identifiers.}

\item{response}{The name of the column containing response values, eg. log2 transformed intensities.}

\item{dose}{The name of the column containing dose values, eg. the treatment concentrations.}

\item{targets}{A character vector that specifies the names of the precursors, peptides or proteins (depending on \code{grouping}) that should be plotted. This can also be \code{"all"} if plots for all curve fits
should be created. If names are provided maximally 20 proteins are plotted at a time, the rest is ignored. If more should be plotted, a mapper over a subsetted data frame should be created.}

\item{unit}{A character vector specifying the unit of the concentration.}

\item{y_axis_name}{A character vector specifying the name of the y-axis of the plot.}

\item{scales}{A character vector that specifies if the scales in faceted plots (if more than one target was provided) should be \code{"free"} or \code{"fixed"}.}

\item{x_axis_scale_log10}{A logical indicating if the x-axis scale should be log10 transformed.}
}
\value{
If \code{targets = "all"} a list containing plots for every unique identifier in the \code{grouping} variable is created. Otherwise a plot for the specified targets is created with maximally 20 facets.
}
\description{
Function for plotting four-parameter dose response curves for each group (precursor, peptide or protein), based on output from \code{fit_drc_4p} function.
}
\examples{
\dontrun{
plot_drc_4p(
data,
grouping = eg_precursor_id,
response = intensity,
dose = concentration,
targets = c("ABCDEFK")
)
}
}
