% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peptide_profile_plot.R
\name{plot_peptide_profiles}
\alias{plot_peptide_profiles}
\title{Peptide abundance profile plot}
\usage{
plot_peptide_profiles(...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function was deprecated due to its name changing to \code{peptide_profile_plot()}.
}
\keyword{internal}
