% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_distribution_plot.R
\name{plot_pval_distribution}
\alias{plot_pval_distribution}
\title{Plot histogram of p-value distribution}
\usage{
plot_pval_distribution(...)
}
\value{
A histogram plot that shows the p-value distribution.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function was deprecated due to its name changing to \code{pval_distribution_plot()}.
}
\keyword{internal}
