% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protti_colors}
\alias{protti_colors}
\title{Color scheme for protti}
\format{
A vector containing 16 colors
}
\source{
Dinas imagination.
}
\usage{
protti_colors
}
\description{
A color scheme for protti that contains 16 colors.
}
\keyword{datasets}
