% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_charge_states.R
\name{qc_charge_states}
\alias{qc_charge_states}
\title{Check charge state distribution}
\usage{
qc_charge_states(
  data,
  sample,
  grouping,
  charge_states,
  intensity = NULL,
  method,
  plot = FALSE
)
}
\arguments{
\item{data}{A dataframe containing at least sample names, peptide or precursor identifiers and missed cleavage counts for each peptide or precursor.}

\item{sample}{The column in the data dataframe containing the sample name.}

\item{grouping}{The column in the data dataframe containing either precursor or peptide identifiers.}

\item{charge_states}{The column in the data dataframe containing the different charge states assigned to the precursor or peptide.}

\item{intensity}{Optional column containing the corresponding intensity values to each peptide or precursor.}

\item{method}{Method used for evaluation. "count" calculates the charge state distribution based on counts of the corresponding peptides or precursors in the charge state group, "intensity" calculates the percentage of precursors or peptides in each charge state group.}

\item{plot}{A logical indicating whether the result should be plotted.}
}
\value{
A data frame that contains the calculated percentage made up by the sum of all peptides or precursors of the corresponding of a specific charge state.
}
\description{
Calculates the charge state distribution for each sample (by count or intensity).
}
\examples{
\dontrun{
qc_charge_states(data, sample, grouping, charge_startes, intensity = NULL, plot = TRUE)
}
}
