% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_charge_states.R
\name{qc_charge_states}
\alias{qc_charge_states}
\title{Check charge state distribution}
\usage{
qc_charge_states(
  data,
  sample,
  grouping,
  charge_states,
  intensity = NULL,
  method,
  plot = FALSE
)
}
\arguments{
\item{data}{A dataframe containing at least sample names, peptide or precursor identifiers and missed cleavage counts for each peptide or precursor.}

\item{sample}{The column in the data dataframe containing the sample name.}

\item{grouping}{The column in the data dataframe containing either precursor or peptide identifiers.}

\item{charge_states}{The column in the data dataframe containing the different charge states assigned to the precursor or peptide.}

\item{intensity}{Column containing the corresponding intensity values for each peptide or precursor. Required when "intensity" is chosen as the method.}

\item{method}{Method used for evaluation. "count" calculates the charge state distribution based on counts of the corresponding peptides or precursors in the charge state group, "intensity" calculates the percentage of precursors or peptides in each charge state group based on the corresponding intensity values.}

\item{plot}{A logical indicating whether the result should be plotted.}
}
\value{
A data frame that contains the calculated percentage made up by the sum of either all counts or intensities of peptides or precursors of the corresponding charge state (depending on which method is chosen).
}
\description{
Calculates the charge state distribution for each sample (by count or intensity).
}
\examples{
\dontrun{
qc_charge_states(
data,
sample = r_file_name,
grouping = pep_stripped_sequence,
charge_states = fg_charge,
intensity = NULL,
method = "count",
plot = TRUE)
}
}
