% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_missed_cleavages.R
\name{qc_missed_cleavages}
\alias{qc_missed_cleavages}
\title{Check missed cleavages}
\usage{
qc_missed_cleavages(
  data,
  sample,
  grouping,
  missed_cleavages,
  intensity = NULL,
  method,
  plot = FALSE
)
}
\arguments{
\item{data}{A dataframe containing at least sample names, peptide or precursor identifiers and missed cleavage counts for each peptide or precursor.}

\item{sample}{The column in the data dataframe containing the sample name.}

\item{grouping}{The column in the data dataframe containing either precursor or peptide identifiers.}

\item{missed_cleavages}{The column in the data dataframe containing the counts of missed cleavages per peptide or precursor.}

\item{intensity}{Optional column containing the corresponding intensity values to each peptide or precursor.}

\item{method}{Method used for evaluation. "count" calculates the percentage of missed cleavages based on counts of the corresponding peptide or precursor, "intensity" calculates the percentage of missed cleavages by intensity of the corresponding peptide or precursor.}

\item{plot}{A logical indicating whether the result should be plotted.}
}
\value{
A data frame that contains the calculated percentage made up by the sum of all peptides or precursors containing the corresponding amount of missed cleavages.
}
\description{
Calculates the percentage of missed cleavages for each sample (by count or intensity).
}
\examples{
\dontrun{
qc_missed_cleavages(data,
sample,
grouping,
missed_cleavages,
intensity,
method = "intensity",
plot = TRUE)
}
}
