% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_missed_cleavages.R
\name{qc_missed_cleavages}
\alias{qc_missed_cleavages}
\title{Check missed cleavages}
\usage{
qc_missed_cleavages(
  data,
  sample,
  grouping,
  missed_cleavages,
  intensity,
  remove_na_intensities = TRUE,
  method,
  plot = FALSE
)
}
\arguments{
\item{data}{A dataframe containing at least sample names, peptide or precursor identifiers and missed cleavage counts for each peptide or precursor.}

\item{sample}{The column in the data dataframe containing the sample name.}

\item{grouping}{The column in the data dataframe containing either precursor or peptide identifiers.}

\item{missed_cleavages}{The column in the data dataframe containing the counts of missed cleavages per peptide or precursor.}

\item{intensity}{Column containing the corresponding intensity values to each peptide or precursor (not log2 transformed).}

\item{remove_na_intensities}{Logical specifying if sample/grouping combinations with intensities that are NA (not quantified IDs) should
be dropped from the data frame for analysis of missed cleavages. Default is TRUE since we are usually
interested in quantifiable peptides.}

\item{method}{Method used for evaluation. "count" calculates the percentage of missed cleavages based on counts of the corresponding peptide or precursor, "intensity" calculates the percentage of missed cleavages by intensity of the corresponding peptide or precursor.}

\item{plot}{A logical indicating whether the result should be plotted.}
}
\value{
A data frame that contains the calculated percentage made up by the sum of all peptides or precursors containing the corresponding amount of missed cleavages.
}
\description{
Calculates the percentage of missed cleavages for each sample (by count or intensity).The default settings remove grouping variables without quantitative information (intensity is NA).
These will not be used for the calculation of missed cleavage percentages.
}
\examples{
\dontrun{
qc_missed_cleavages(
data,
sample = r_file_name,
grouping = pep_stripped_sequence,
missed_cleavages = pep_nr_of_missed_cleavages,
intensity = fg_quantity,
method = "intensity",
plot = TRUE)
}
}
