% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_pca.R
\name{qc_pca}
\alias{qc_pca}
\title{Plot principal component analysis}
\usage{
qc_pca(data, sample, grouping, intensity, condition, digestion = NULL)
}
\arguments{
\item{data}{A dataframe containing  sample names, peptide or precursor identifiers, corresponding intensities and a condition column indicating e.g. the treatment.}

\item{sample}{The column in the data dataframe containing the sample name.}

\item{grouping}{The column in the data dataframe containing either precursor or peptide identifiers.}

\item{intensity}{Column containing the corresponding intensity values for each peptide or precursor.}

\item{condition}{Column indicating the treatment or condition for each sample.}

\item{digestion}{Optional column indicating the mode of digestion (limited proteolysis or tryptic digest).}
}
\value{
A plotted principal component analysis showing PC1 and PC2
}
\description{
Plots a principal component analysis based on peptide or precursor intensities.
}
\examples{
\dontrun{
qc_PCA(
data,
sample = r_file_name,
grouping = eg_precursor_id,
intensity = normalised_intensity_log2,
condition = r_condition
)
}

}
