% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_identified_by_x.R
\name{replace_identified_by_x}
\alias{replace_identified_by_x}
\title{Replace identified positions in protein sequence by "x"}
\usage{
replace_identified_by_x(sequence, positions_start, positions_end)
}
\arguments{
\item{sequence}{Sequence a character vector that contains the protein sequence.}

\item{positions_start}{A vector of start positions of the identified peptides.}

\item{positions_end}{A vector of end positions of the identified peptides.}
}
\value{
A modified protein sequence with each identified position replaced by "x"
}
\description{
Helper function for the calculation of sequence coverage, replaces identified positions with an "x" within the protein sequence.
}
\examples{
\dontrun{
replace_identified_by_x(
sequence = c("AEFGPEEAAVS"),
positions_start = c(1, 5, 8),
positions_end = c(3, 6, 11)
)
}

}
