% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_identified_by_x.R
\name{replace_identified_by_x}
\alias{replace_identified_by_x}
\title{Replace identified positions in protein sequence by "x"}
\usage{
replace_identified_by_x(sequence, positions_start, positions_end)
}
\arguments{
\item{sequence}{a character value that contains the protein sequence.}

\item{positions_start}{a numeric vector of start positions of the identified peptides.}

\item{positions_end}{a numeric vector of end positions of the identified peptides.}
}
\value{
A character vector that contains the modified protein sequence with each identified
position replaced by "x".
}
\description{
Helper function for the calculation of sequence coverage, replaces identified positions with an
"x" within the protein sequence.
}
