% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_query.R
\name{try_query}
\alias{try_query}
\title{Query from url}
\usage{
try_query(url, max_tries = 5, silent = TRUE)
}
\arguments{
\item{url}{Url of website that contains the table that should be downloaded.}

\item{max_tries}{Number of times the function tries to download the data in case an error occures.}

\item{silent}{logical, if TRUE no individual messages are printed after each try that failed.}
}
\value{
A data frame that contains the table from the url.
}
\description{
Downloads data from url using the \code{fread} function from the \code{data.table} package. If an error occures during the query (for example due to no connection) the function waits 3 seconds and tries again.
}
\examples{
\dontrun{
try_query("http://www.uniprot.org/uniprot/?query=id:P36578&format=tab")
}
}
