% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_protti.R
\name{volcano_protti}
\alias{volcano_protti}
\title{Volcano plot}
\usage{
volcano_protti(
  data,
  grouping,
  foldchange,
  p_value,
  method,
  protein_identifier = NULL,
  target = NULL,
  title = "Volcano plot",
  x_axis_label = "log2(fold change)",
  foldchange_cutoff = 1,
  p_value_cutoff = 0.01
)
}
\arguments{
\item{data}{Dataframe containing at least sample names, peptide or precursor identifiers and missed cleavage counts for each peptide or precursor.}

\item{grouping}{Column in the data dataframe containing either precursor or peptide identifiers.}

\item{foldchange}{Column in the dataframe containing the fold changes (ideally log2 transformed) between two conditions}

\item{p_value}{Column containing the p-value for the corresponding fold changes. P-value is ideally adjusted using e.g. Benjamini-Hochberg correction.}

\item{method}{Method used for the plot. \code{method = "target"} highlights your protein of interest in the volcano plot, \code{method = "significant"} highlights all significantly changing entities.}

\item{protein_identifier}{Optional column required for \code{method = "target"}, contains protein identifiers (e.g. UniProt IDs).}

\item{target}{Optional argument required for \code{method = "target"}, protein identifier for your protein of interest.}

\item{title}{Optional argument specifying the title of the volcano plot. Default is "Volcano plot".}

\item{x_axis_label}{Optional argument specifying the x-axis label. Default is "log2(fold change)".}

\item{foldchange_cutoff}{Optional argument specifying the fold change cutoff used for assessing whether changes are significant. As a log2 scale should be used, the default value is 1.}

\item{p_value_cutoff}{Optional argument specifying the p-value cutoff used for assessing significance of changes. Default is 0.01.}
}
\value{
Depending on the method used a volcano plot with either highlighted target protein (\code{method = "target"}) or highlighted significant proteins (\code{method = "significant"}).
}
\description{
Plots a volcano plot for the given input
}
\examples{
\dontrun{
volcano_protti(
data,
grouping = pep_stripped_sequence,
foldchange = Log2FC,
p_value = p_value,
method = "target",
target = Q9Y6K9,
title = "Finding Nemo",
x_axis_label = "log2(fold change) treated vs untreated",
foldchange_cutoff = 0.5,
p_value_cutoff = 0.05
)
}
}
